Starts consolidating form hooks
