+++ src/components/forms/ArticleManagementForm.jsx
@@ -1,6 +1,7 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import { withRouter } from 'react-router-dom'
-import axios from '../../helpers/axiosService'
+
+import { useArticleManagementHook } from './formHooks'
 
 const ArticleManagementForm = (props) => {
   const [
@@ -72,32 +73,3 @@ const ArticleManagementForm = (props) => {
 }
 
 export default withRouter(ArticleManagementForm)
-
-const useArticleManagementHook = (props) => {
-  const [formData, setFormData] = useState({})
-  const isCreate = props.match.params.action === 'create'
-  useEffect(() => {
-    !isCreate && setFormData(props.location.state.article)
-  }, [])
-
-  const onInputChangeHandler = (e) => {
-    setFormData({ ...formData, [e.target.name]: e.target.value })
-  }
-  const submitArticle = async (e) => {
-    e.preventDefault()
-    const payload = { article: formData }
-
-    try {
-      const response = isCreate
-        ? await axios.createArticle(payload)
-        : await axios.updateArticle(props.location.state.article.slug, payload)
-      props.history.push({
-        pathname: `/article/${response.data.article.slug}`,
-      })
-    } catch (error) {
-      console.error(error)
-    }
-  }
-
-  return [formData, onInputChangeHandler, submitArticle]
-}
