+++ src/components/forms/Settings.jsx
@@ -1,17 +1,15 @@
-import React, { useState, useEffect, useContext } from 'react'
+import React from 'react'
 import { withRouter } from 'react-router-dom'
 
-import axios from '../../helpers/axiosService'
-import { AppContext } from '../../AppContext'
-import { deleteToken } from '../../helpers/localStorage'
+import useFormsHook from './formHooks'
 
 const Settings = (props) => {
-  const [
+  const {
     formData,
     updateUser,
     logoutUser,
     onInputChangeHandler,
-  ] = useSettingsHook(props)
+  } = useFormsHook(props)
   return (
     <div className="settings-page">
       <div className="container page">
@@ -89,40 +87,3 @@ const Settings = (props) => {
 }
 
 export default withRouter(Settings)
-
-const useSettingsHook = (props) => {
-  const [state, setState] = useContext(AppContext)
-  const [formData, setFormData] = useState({})
-  useEffect(() => {
-    setFormData(state.currentUser)
-  }, [state])
-  const onInputChangeHandler = (e) => {
-    setFormData({ ...formData, [e.target.name]: e.target.value })
-  }
-  const updateUser = async (e) => {
-    e.preventDefault()
-    const payload = { user: formData }
-    try {
-      const response = await axios.updateUser(payload)
-      const currentUser = {
-        isSignedIn: true,
-        username: response.data.user.username,
-      }
-      setState((state) => ({ ...state, currentUser: currentUser }))
-      props.history.push({
-        pathname: '/',
-      })
-    } catch (error) {
-      console.error(error)
-    }
-  }
-  const logoutUser = () => {
-    deleteToken()
-    setState((state) => ({ ...state, currentUser: { isSignedIn: false } }))
-    props.history.push({
-      pathname: '/',
-    })
-  }
-
-  return [formData, updateUser, logoutUser, onInputChangeHandler]
-}
